Fix snake being broken after 255 cherries

+++ src/cell.h
@@ -31,7 +31,7 @@ struct cell_s;
 typedef struct cell_s {
     bool           has_cherry;
     bool           has_snake;
-    uint8_t        snake_counter;
+    uint16_t       snake_counter;
     direction_t    snake_direction;
     struct cell_s *previous_snake_cell;
 } cell_t;
